+++ src/consoleCmdSet.h
@@ -14,11 +14,11 @@ extern void consoleCmdSetUpdateCmds();
 char *consoleCmdGetBackground();                      // [SCRIPT_NAME: get_background]
 void consoleCmdSetBackground(char *color);            // [SCRIPT_NAME: set_background]
 
-float consoleCmdGetCampos(int index);                 // [SCRIPT_NAME: get_campos]
-void consoleCmdSetCampos(int index, float value);     // [SCRIPT_NAME: set_campos]
+double consoleCmdGetCampos(int index);                 // [SCRIPT_NAME: get_campos]
+void consoleCmdSetCampos(int index, double value);     // [SCRIPT_NAME: set_campos]
 
-float consoleCmdGetCamposl(int index);                // [SCRIPT_NAME: get_camposl]
-void consoleCmdSetCamposl(int index, float value);    // [SCRIPT_NAME: set_camposl]
+double consoleCmdGetCamposl(int index);                // [SCRIPT_NAME: get_camposl]
+void consoleCmdSetCamposl(int index, double value);    // [SCRIPT_NAME: set_camposl]
 
 bool consoleCmdGetConvexhull();                       // [SCRIPT_NAME: get_convexhull]
 void consoleCmdSetConvexhull(bool value);             // [SCRIPT_NAME: set_convexhull]
@@ -26,8 +26,8 @@ void consoleCmdSetConvexhull(bool value);             // [SCRIPT_NAME: set_conve
 int consoleCmdGetDimen();                             // [SCRIPT_NAME: get_dimen]
 void consoleCmdSetDimen(bool value);                  // [SCRIPT_NAME: set_dimen]
 
-float consoleCmdGetEdgesize();                        // [SCRIPT_NAME: get_edgesize]
-void consoleCmdSetEdgesize(float value);              // [SCRIPT_NAME: set_edgesize]
+double consoleCmdGetEdgesize();                        // [SCRIPT_NAME: get_edgesize]
+void consoleCmdSetEdgesize(double value);              // [SCRIPT_NAME: set_edgesize]
 
 char *consoleCmdGetFacecolor();                       // [SCRIPT_NAME: get_facecolor]
 void consoleCmdSetFacecolor(char *color);             // [SCRIPT_NAME: set_facecolor]
@@ -41,8 +41,8 @@ void consoleCmdSetHistory(int value);                 // [SCRIPT_NAME: set_histo
 int consoleCmdGetMaxfps();                            // [SCRIPT_NAME: get_maxfps]
 void consoleCmdSetMaxfps(int value);                  // [SCRIPT_NAME: set_maxfps]
 
-float consoleCmdGetMousesens();                       // [SCRIPT_NAME: get_mousesens]
-void consoleCmdSetMousesens(float value);             // [SCRIPT_NAME: set_mousesens]
+double consoleCmdGetMousesens();                       // [SCRIPT_NAME: get_mousesens]
+void consoleCmdSetMousesens(double value);             // [SCRIPT_NAME: set_mousesens]
 
 bool consoleCmdGetPyexpr();                           // [SCRIPT_NAME: get_pyexpr]
 void consoleCmdSetPyexpr(bool value);                 // [SCRIPT_NAME: set_pyexpr]
@@ -50,17 +50,17 @@ void consoleCmdSetPyexpr(bool value);                 // [SCRIPT_NAME: set_pyexp
 char *consoleCmdGetSelvertcolor();                    // [SCRIPT_NAME: get_selvertcolor]
 void consoleCmdSetSelvertcolor(char *color);          // [SCRIPT_NAME: set_selvertcolor]
 
-float consoleCmdGetSelvertsize();                     // [SCRIPT_NAME: get_selvertsize]
-void consoleCmdSetSelvertsize(float value);           // [SCRIPT_NAME: set_selvertsize]
+double consoleCmdGetSelvertsize();                     // [SCRIPT_NAME: get_selvertsize]
+void consoleCmdSetSelvertsize(double value);           // [SCRIPT_NAME: set_selvertsize]
 
 char *consoleCmdGetSpacecolor(int index);             // [SCRIPT_NAME: get_spacecolor]
 void consoleCmdSetSpacecolor(int index, char *color); // [SCRIPT_NAME: set_spacecolor]
 
-float consoleCmdGetSpeed();                           // [SCRIPT_NAME: get_speed]
-void consoleCmdSetSpeed(float value);                 // [SCRIPT_NAME: set_speed]
+double consoleCmdGetSpeed();                           // [SCRIPT_NAME: get_speed]
+void consoleCmdSetSpeed(double value);                 // [SCRIPT_NAME: set_speed]
 
 bool consoleCmdGetStdoutpyexpr();                     // [SCRIPT_NAME: get_stdoutpyexpr]
 void consoleCmdSetStdoutpyexpr(bool value);           // [SCRIPT_NAME: set_stdoutpyexpr]
 
-float consoleCmdGetVertsize();                        // [SCRIPT_NAME: get_vertsize]
-void consoleCmdSetVertsize(float value);              // [SCRIPT_NAME: set_vertsize]
+double consoleCmdGetVertsize();                        // [SCRIPT_NAME: get_vertsize]
+void consoleCmdSetVertsize(double value);              // [SCRIPT_NAME: set_vertsize]
